+++ HISTORY.rst
@@ -2,6 +2,11 @@
 
 History
 -------
+0.1.3 (2016-01-26)
+++++++++++++++++++
+
+* Fix setting certificate.
+* Add method DjangoTelegramBot.getBot(); get bot instance by token or username.
 
 0.1.2 (2016-01-26)
 ++++++++++++++++++